Break apart hostmap.go
